Add target to run KBoot in QEMU
If a test script is present for the current configuration in utilities/test,
a 'qemu' target will be added that runs that script. The script generates a
bootable image and runs it in QEMU. Later this target will also build the
KBoot test kernel and add that to the image.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>