Add initial UEFI platform code (currently 64-bit x86 only)
Adds the initial build infrastructure for UEFI platforms, and an efi-amd64
configuration to build the loader to run under UEFI on 64-bit x86 systems.
Since UEFI supports multiple architectures, the UEFI platform is split
into generic and architecture dependent code.

Currently a bootx64.efi binary will be built which can be run under QEMU
with OVMF, and will print a message on the serial port and then halt.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>