test: Allow OVMF non-volatile variables to be saved across QEMU runs
Modified the EFI test script to make a local copy of the OVMF binary and
load it in QEMU using the -pflash option rather than -bios. This means
that OVMF will save it's non-volatile variables in the local copy of the
OVMF binary and persist across QEMU runs. Also changed the default
settings to not output the primary console to the serial port - KBoot
treats the on-screen and serial consoles differently.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>