x86: implement exception handling
Implement exception handling on x86 for both BIOS and EFI platforms. This
does add some complexity for EFI (we have to wrap all EFI calls to restore
the original GDT/IDT/segment registers that were in place when the loader
was started), however it gets us useful information on exceptions that
occur in the loader rather than a triple fault - the EFI implementations I
have tried do not seem to do any exception handling.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>