Add virtual to physical address conversion functions
On some architectures (e.g. MIPS) the loader does not run in the physical
address space or in an identity mapped virtual address space. Therefore
we need to have functions that can be used to convert between physical
addresses and virtual addresses that the loader can access. At the moment
we just use an architecture defined offset to add to a physical address
to get a virtual address. This is sufficient for MIPS, which will just
define the offset as the start of KSEG0.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>