EFI: Implement memory allocation functions
On EFI, we don't use the generic memory management code, as when in boot
services mode the firmware is in control of the memory map, and it is
possible for the memory map to change underneath the boot loader. Since
the AllocatePages boot service cannot implement all the functionality
of memory_alloc() (it does not implement alignment or minimum address
constraints), we instead implement memory_alloc() by getting the current
memory map each time it is called, scanning it for a suitable free range,
and then allocating the exact address we want with AllocatePages.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>