Implement device manager and parts of the disk device layer
The device manager maintains a list of devices and allows them to be
looked up by name (and soon, by the UUID/label of the filesystem on the
device), and provides an API to read from devices.

The disk device layer manages block devices (disks), and will soon
handle partition maps.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>