Stop using global state in console implementations
Add private data pointers to console_t for console implementations to store
internal state, and stop using global state in the console implementations.
With the UI implementation we need to add a bunch more internal state in the
console drivers and it's nicer to bundle everything into a state structure
rather than storing it globally. Also would allow the same console driver to
be used for multiple consoles (e.g. multiple serial ports).

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>