Cleanups/improvements for filesystem code
Rather than having iterate() give a handle to each entry found to the
callback function, instead pass it an fs_entry_t (embedded inside an FS-
specific structure that contains information needed to access the file)
that can then be passed to a new open_entry() function to open a handle
to it.

The reason this is done is because on some filesystems (e.g. ext2) opening
a handle to a file is quite heavyweight and may not actually need to be
done (for example, in the previous implementation, opening a file in a
large directory would result in the inodes for every entry preceding the
one actually being opened being read in).

This also removes the need for reference counting on fs_handle_t - the
reason for it being there was for an iterate() callback such as the one
used in implementing fs_open() to cause a handle to be retained after it
returns. Since the handle is now manually opened by the callback,
reference counting becomes unnecessary.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>