Improve lsdev output
Extend the device identify() callback to allow a long, multi-line string
to be returned, which will be printed by lsdev when getting information
on a specific device. For disks this prints the block size and block
count, rather than tagging them on the end of the basic identification
string (which makes that string quite long).

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>