Replace uses of CONFIG_64BIT with __LP64__
CONFIG_64BIT is fixed by the configuration regardless of the actual
compilation mode. The KBoot test kernel will reuse the main loader
headers, however itcan be compiled in a different mode to the main
loader, e.g. we compile both a 32- and 64-bit test kernel on x86. To
make the headers reusable, use __LP64__ instead which is set by the
compiler based on the compilation mode.

CONFIG_{32,64}BIT are still retained in the configuration for use by
the build system.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>