Physical memory allocator changes
By default memory_alloc() will now call boot_error() if there is not
enough memory available to satisfy a request. This is because in the
majority of cases this will be a boot failure, so we may as well error
in a single place rather than scattering error messages around at almost
every call to memory_alloc(). A new flag MEMORY_ALLOC_CAN_FAIL is added
to allow the call to return NULL if it fails should this behaviour be
required.

Additionally, avoid allocating below 1MB on x86 systems unless a caller
explicitly requests this, as firmwares like to do funny things with
low memory.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>