Improve configuration error handling
First of all, remove the need to manually prefix the string passed
to config_error() with the command name. Just store the name of the
currently executing command and add it automatically.

Secondly, add the ability to set an error callback function for
config_error(). This replaces the hardcoded logic to handle errors
when in the shell, and is used by the menu error handling.

Finally, errors that occur in the commands within an entry no longer
cause the entire configuration file load to fail. Instead, we buffer
up the error string passed to config_error(), and display that with
boot_error() when the user actually tries to boot the entry.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>