Add command to set video mode used by the loader
Add a command which immediately changes the video mode being used by
the loader. Placing it globally in the configuration file will cause
the menu interface to be displayed in the specified mode. The
video_mode environment variable only selects the video mode which will
be set when an OS is actually loaded.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>