test: Implement some basic memory management/mapping
As mentioned in the previous commit using a fixed size mapping of physical
memory is not sufficient on 32-bit platforms as the loader can allocate
anywhere within the low 4GB of memory. To avoid this problem, implement
some basic memory management/mapping code for 32-bit platforms to map in
physical memory. For IA32, this uses the recursive pagetable mapping set
up by the loader. 64-bit platforms still use a fixed mapping of physical
memory, so between 32-bit and 64-bit we now get test coverage of both
KBOOT_ITAG_MAPPING and the recursive page table mapping.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>