BIOS: Improve Multiboot support
When the loader is booted via Multiboot and passed modules, the modules
will be used to populate a virtual filesystem. If no configuration file
is loaded as a module, one will be automatically generated that will load
the first module as a KBoot kernel, and any remaining modules as modules
for the kernel. Options for the kernel can be specified on the module
command line, which will be turned into appropriate set commands in the
generated configuration.

One use of this is to allow a KBoot kernel to be booted easily via QEMU's
-kernel command line option, for example:

  $ qemu -kernel loader -initrd "kernel.elf video_mode=\"lfb:800x600x32\",module.bin"

This will load kernel.elf, with an 800x600x32 video mode, and load
module.bin as a module.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>