Allow bringing up the config menu when the OS to boot is defined by the root environment
When the OS to boot is defined by the root environment rather than a
menu entry, pressing F8 will now bring up the configuration menu for
the OS (if any). F2 will also bring up the shell.

Also change the Multiboot automatic configuration generation to put all
commands at the top level to make use of this, so we don't have to
define a dummy "KBoot Kernel" menu entry just so that the configuration
interface can be brought up.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>