Major refactor of console code, add proper serial console support
Consoles are now named, and we can have any number of consoles, rather
than just the main and debug consoles. Remove ui_console and
config_console and their respective printf variants, now we just have
current_console which is the console that all UI and the shell goes to.
The current console can be set via a config command.

Also added proper serial console support. A generic serial layer has
been added, which handles things like port configuration (a config
command has been added to set the port configuration), escape code
parsing, etc. The current console can be set to a serial console, which
allows the shell to be used over a serial console. At the moment we
don't have UI support on the serial port, but this can be added later.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>