Add current_time() function
Add function to get a monotonically increasing internal time value. As a
side effect we no longer have a custom delay implementation for EFI, we
just use the TSC as we do on BIOS. This is because there's no easy way
to get a time like we want using EFI services. I tried getting the real
time, which can support high precision, but everywhere I've tested it
only has second precision. This *might* cause a problem somewhere because
we use the PIT for calibration of the TSC, which may screw with the way
the firmware may have set the PIT up, but for now it seems to be OK.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>