Fixed a few bugs
Remove the check for consistency between the terminal's reported cursor
position and the internal cursor position. I'm fairly sure everything's
correct now, and it can be problematic if there is a lot of input escape
codes being sent, which will confuse the code to get the reported cursor
position.

Also fixed incorrect colour usage in the shell prompt.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>