EFI: Add fixed iPXE ROMs for PXE QEMU test target
The iPXE EFI ROMs currently included with QEMU have a bug. iPXE registers
its own load file protocol on network device handles, however doing so
causes EDK to fail to register the PXE base code protocol on the handle,
as it tries to do so at the same time as installing its own load file
protocol). This means that the PXE base code protocol is unavailable to
the loaded image.

The ROMs added by this commit are built with the following patches:

  http://lists.ipxe.org/pipermail/ipxe-devel/2015-March/004008.html
  http://lists.ipxe.org/pipermail/ipxe-devel/2015-March/004009.html

The latter of these disables iPXE's own load file protocol and allows
PXE boot to work properly with OVMF in QEMU.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>