x86: Add Linux kernel loading support
Add x86 Linux kernel loading support. On BIOS we use the 32-bit boot
protocol, on EFI we use the EFI handover protocol (which requires a 3.6
or later kernel).

There's still a couple of TODOs, in particular I'm seeing some weird
behaviour if booting with a VGA console that I need to investigate.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>