Add 'debug' command to set the debug console
May be useful for instance to redirect the debug log to the primary
console to debug problems on systems where a serial console is not
available. Also added an active count to console_t to track how many
users of a console there are, so we can avoid reinitializing a console
that's already initialized (or similarly deinitializing one that is
still in use) when setting both current and debug consoles to the same
console.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>