ext2boot: Implement path resolution to replace hardcoded inode number
This is much less fragile than hardcoding the loader inode number, which
could change under certain circumstances. I considered making use of the
EXT2_BOOT_LOADER_INO inode (which Linux now provides a mechanism to
access), but this would mean that updating the boot loader would require
an extra step to copy it into that inode.

With path resolution, updating the loader is a simple matter of replacing
the binary. No need to worry about keeping the inode number the same.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>