Add installation utility
This utility supports installation to a directory, device, or disk image.
When installing to a directory, it is capable of copying the boot loader
binary to that directory, determining the device containing the directory,
and installing the boot sector to it. When installing to a device or
image, the boot loader binary must already exist on the filesystem, and it
can install the boot sector with a user-supplied binary path.

Only implemented for BIOS at the moment. Not well tested, use at your own
risk.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>