Fix bug where partition offset in disk_device_t could be overwritten
Partition information was in a union with some information for raw disks
to save a little space. However, if the disk probe code identifies that
a partition has a partition map type (even if no partitions were found),
this could cause partition_map to be set for the partition, which would
overwrite its offset. This can in fact happen for any bootable partition
(without a recognised FS, as FSes are probed before partition maps) as
they will have the MBR signature in them.

Since it can in future be possible for partitions to legitimately contain
their own partition map (e.g. BSD disklabels), fix this by un-unioning
the members in disk_device_t.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>