Implement a basic graphical boot menu
The menu currently allows an icon image to be assigned to each entry, and
will lay them out horizontally on the screen. A selection image must be
specified, which will be drawn underneath icons to indicate the currently
selected entry. The same inputs as the text boot menu are supported, for
instance pressing F1 on an entry will bring up the usual configuration
text UI.

Still to do are add an indicator of the remaining timeout, and support
for background images in addition to just a solid colour.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>