EFI: Detect boot directory and look for configuration file there
Set the current directory upon boot to the directory containing the boot
loader image, and then look for "kboot.cfg" in there in preference to
all other default config paths.

This means that we can always place the configuration file in the same
directory as the boot loader and find it correctly.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>