Add {malloc,free}_large() functions for large allocations
The internal heap is a small, fixed-size area which makes it unsuitable
for larger allocations. In some places large allocations are made using
memory_alloc() with the type set as MEMORY_TYPE_INTERNAL. Add a new set
of convenience functions which call memory_alloc() with the right flags
for large allocations, and track allocation sizes to avoid having to
recalculate the size when freeing.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>