EFI/amd64: Fix stack alignment when jumping to C code
On both the SysV and MS x86_64 ABIs, RSP + 8 must be a multiple of 16
upon entry to a function. The startup code did not correctly implement
this requirement. This misalignment propagates through to EFI calls and
loaded EFI images.

This was causing the Windows boot loader to crash when chain loaded by
KBoot. With this commit KBoot can now chain load Windows on EFI.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>