Add FS_OPEN_DECOMPRESS flag for fs_open(), don't decompress unless set
We don't always want to automatically decompress compressed files, so
add a flag to control when it is done. Currently the KBoot and Multiboot
loaders use it to decompress the kernel/modules.

The motivation for doing this was that KBoot was decompressing the
initramfs for my Linux system on boot (despite the kernel being capable
of doing it itself) which was significantly slowing down boot.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>