ext2boot: Bake in a partition LBA as a fallback if no information provided by MBR
We were relying on the MBR passing partition information when booting from
a partition to get its LBA. While the standard MBR does provide this
information, some MBR implementations do not, and neither does Apple's
BIOS emulation. Therefore, bake in a partition LBA when installing and
fall back on this if not provided with valid partition information. The
provided information is still preferred when valid because using that
allows the boot sector/partition to be moved around without having to
update it.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>