Improvements to the GUI menu
Background images are now supported. A solid colour is also supported for
the selection, which is filled in the area behind the entry icon when it
is selected. The behaviour of selection images has changed: rather than
being drawn underneath the entry icon, it is drawn centred on the centre
of the icon. This allows, for example, a glow effect to be drawn around
circular icons.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>