Avoid usage of CONTENT_HEIGHT when it is not valid
In ui_list_insert() CONTENT_HEIGHT was being used to set the current list
offset so that the selected entry is visible. However, CONTENT_HEIGHT is
only valid inside ui_display(). This was causing an internal_error() in
certain cases. Set the list offset at the top of ui_list_render() if the
currently selected entry is off-screen instead.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>