ext2boot: Fix handling of directories spanning multiple blocks
At the start of a block iteration, the temporary buffer offset must be
reset. Its value is compared with the block size to determine if the
end of a block is reached, so after the first block that test would
always decide that it was at the end of the block, thus skipping most
of the directory contents past the first block. Also, on very large
directories, it could overflow if not reset.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>