Fix assert failure if video_mode is set to a shortened mode string
When creating the video mode chooser for the configuration UI, the mode
string in the video_mode variable must exactly match an existing mode
for the selected entry to be picked correctly. If the video_mode value
specified in the configuration is a shortened string, e.g. "vga" as
opposed to "vga:80x25", it would remain set to that causing an assertion
failure in the UI when opening the configuration menu, because the
selected entry would be NULL. Fix by setting video_mode to the actual
mode a shortened string resolves to in video_env_init().

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>