Improve GUI menu drawing to reduce visible glitches
When changing the selected entry, there can be noticable glitches while
redrawing the old and new entries. This is because we redraw in 3 steps
(draw the background in the icon area, draw the selection image/colour,
draw the icon). The framebuffer is updated on each of these steps, and
the intermediate results can become visible on screen.

This commit improves things to avoid causing glitches by performing the
whole redraw in the backbuffer, and only updating the real framebuffer
once the whole redraw is complete.