Make ARM64/DT loader relocatable, implement early init code
The DT platform is intended to be usable across any platform which provides
a device tree, therefore we have no guarantees about where the loader image
will be loaded in physical memory. This means the loader must be relocatable.

Relocations are applied during early init, like we do for EFI.

This now gets as far as calling into C init code.