Add KBOOT_TAG_SERIAL tag to pass serial port configuration to the kernel
This will allow a kernel to continue using the same serial port that was
configured for use in KBoot (if any), for example as a debug console. This is
also useful for getting easy early access to a serial console before it can be
discovered via platform-specific mechanisms (e.g. device tree).