Add support for specifying cacheability for KBoot mappings
On x86 we never cared about this because the MTRRs are typically set up to set
the right cache behaviour for memory-mapped devices.

For ARM we need to set up cacheability of mappings ourselves. Kiwi and the
test kernel use a direct mapping of physical memory set up through a KBoot
mapping tag to access physical memory, including for device access. Device
access needs to disable caching so caching on this mapping needs to be
configured accordingly.

Therefore add a field to KBOOT_ITAG_MAPPING to specify cacheability of the
mapping. This requires a protocol version bump so we can detect whether or not
the flag will be there.

Also, set appropriate cacheability for some other mappings that KBoot makes
automatically: the framebuffer is set as writethrough, while the kernel entry
trampoline mappings are set as uncached so that the kernel does not need to
worry about clearing them from the caches.